reorganize components by page/screen and refactor background image
