+++ src/App.scss
@@ -1,0 +1,14 @@
+* {
+  padding: 0px;
+}
+
+body {
+
+  .App {
+    background-image: url("./GlobalAssets/styles/Black-Texture.jpg");
+    background-repeat: no-repeat;
+    background-size: cover;
+  }
+  
+}
+
