--- src/Components/app/App.scss
@@ -1,16 +1,0 @@
-* {
-  padding: 0px;
-}
-
-body {
-  background-image: url("../login/Black-Texture.jpg");
-  background-repeat: no-repeat;
-  background-size: cover;
-
-  .App {
-    background-image: url("../login/Black-Texture.jpg");
-    background-repeat: no-repeat;
-    background-size: cover;
-  }
-}
-
