--- src/Components/card/Card.scss
@@ -1,19 +1,0 @@
-@import "../../constants/constants.scss";
-
-.quote-of-the-day {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  border-radius: $borderRadius;
-  background-color: $primaryMono1;
-  box-shadow: 10px 12px 15px rgba(0, 0, 0, 0.35);
-  h2 {
-    font-size: 2em;
-    margin-top: 7%;
-  }
-  p {
-    margin: 5%;
-    font-size: 1.8em;
-  }
-}
