--- src/Components/flashcardPage /components/updateFlashcard/UpdateFlashcard.scss
@@ -1,60 +1,0 @@
-@import "../../../../constants/constants.scss";
-
-.MuiModal-root {
-  display: grid;
-  grid-template-columns: 1fr;
-  .MuiBox-root {
-    display: grid;
-    grid-template-columns: 1fr;
-    place-self: center;
-    height: 80vh;
-    width: 35vw;
-    border-radius: $borderRadius;
-    background-color: $primary;
-
-    .MuiFormControl-root {
-      height: 30vh;
-      background-color: $primaryTint10;
-      margin: 2%;
-      label {
-        margin: 2%;
-        color: white;
-        font-size: 4vh;
-      }
-      .MuiInput-root {
-        margin: 0px;
-        height: 100%;
-        .MuiInput-input {
-          margin: 3%;
-          color: white;
-          font-size: 1.2rem;
-        }
-      }
-    }
-    .update-form-header {
-      display: flex;
-      justify-content: space-between;
-      .update-flashcards-rating-container {
-        margin-right: 0.5vw;
-        align-self: center;
-        border-radius: $borderRadiusInputs;
-        legend {
-          font-size: 2.2vh;
-          font-weight: 900;
-          color: white;
-        }
-        .MuiRating-root {
-          label {
-            height: 3vh;
-          }
-        }
-      }
-      h3 {
-        color: white;
-        align-self: center;
-        margin-left: 0.5vw;
-        font-size: 2.5vw;
-      }
-    }
-  }
-}
