--- src/Components/flashcardPage /flashcardList/FlashcardList.scss
@@ -1,47 +1,0 @@
-@import "../../../constants/constants.scss";
-
-.flashcard-list-container {
-  margin: 0 auto;
-  width: 90vw;
-  background-color: $primaryComplement;
-  color: white;
-  border-radius: $borderRadius;
-  display: grid;
-  .MuiButtonBase-root {
-    place-self: center;
-  }
-  .cards-list {
-    margin: 10px;
-    .MuiGrid-container {
-      .MuiRating-root {
-        width: 5vw;
-      }
-    }
-  }
-}
-
-.deck-list-header {
-  display: flex;
-  justify-content: space-between;
-  h3 {
-    margin-top: 1vh;
-    margin-left: 1vh;
-  }
-  .deck-list-filter {
-    margin-top: 1vh;
-    margin-right: 1vh;
-  }
-}
-
-.cards-list {
-  display: grid;
-  gap: 1vh;
-}
-
-.MuiGrid-root {
-  background-color: $primaryTriadic2;
-  .MuiButton-root {
-    max-height: 4vh;
-    place-self: center;
-  }
-}
