--- src/Components/nav/Nav.scss
@@ -1,51 +1,0 @@
-@import "../../constants/constants.scss";
-
-.nav {
-  background-color: rgba(129, 129, 129, 0.25);
-  backdrop-filter: blur(5px);
-  height: 7vh;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1%;
-  padding-right: 5%;
-  padding-left: 5%;
-  box-shadow: 0px 15px 15px rgba(0, 0, 0);
-
-  h2 {
-    font-size: 7vh;
-  }
-}
-
-.nav-button {
-    background-color: white;
-    width: 50%;
-    height: 40%;
-    border-radius: 15px;
-
-    &:hover{
-      cursor: pointer;
-      background-color: #bababa;
-      
-    }
-  }
-
-.appName {
-    color: white;
-    text-decoration: none;
-}
-
-.css-1s2u09g-control {
-  display: flex;
-  align-items: center;
-}
-
-.css-hv18hu-menu {
-  width: 20%;
-}
-
-.css-tnip6t-menu {
-  z-index: 1;
-}
-
-
