--- src/Components/stats-chart/statsChart.scss
@@ -1,2 +1,0 @@
-@import '../../constants/constants.scss';
-
