+++ src/GlobalAssets/decks/Decks.scss
@@ -1,0 +1,102 @@
+@import "../../GlobalAssets/styles/constants.scss";
+
+.dashboard-deck-container {
+  border-radius: $borderRadius;
+  background-color: $primaryMono1;
+  box-shadow: 2px 2px 10px $primaryTint10;
+}
+.dashboard-deck-list,
+.flashcards-deck-list {
+  border-radius: $borderRadius;
+  background-color: $primaryMono1;
+  display: flex;
+  .nav-link {
+    border-radius: $borderRadiusInputs;
+    box-shadow: 2px 2px 10px $primaryTint10;
+    background-color: $primaryTriadic2;
+    text-decoration: none;
+    color: white;
+    font-size: 1em;
+  }
+  .selected {
+    background-color: $primaryMono2;
+    color: black;
+  }
+  a:hover {
+    background-color: $primaryMono2;
+    color: black;
+  }
+}
+
+.flashcards-deck-container {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-bottom: 5%;
+  box-shadow: rgba(0, 0, 0, 0.25);
+  .flashcards-deck-list {
+    width: 80%;
+    justify-content: space-evenly;
+    box-shadow: 10px 12px 15px rgba(0, 0, 0, 0.35);
+    .nav-link {
+      width: 25%;
+      margin: 2vh;
+      font-size: 3vh;
+      text-align: center;
+      h3 {
+        font-size: 1em;
+        margin: 1%;
+      }
+
+      p {
+        text-align: center;
+        margin-bottom: 0.5vh;
+        margin-left: 0.5vh;
+      }
+    }
+  }
+}
+
+.dashboard-deck-container {
+  box-shadow: 10px 12px 15px rgba(0, 0, 0, 0.35);
+  height: 100%;
+  display: flex;
+  justify-content: center;
+
+  .dashboard-deck-list {
+    list-style-type: none;
+    height: inherit;
+    width: 80%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+
+    h2 {
+      font-size: 2em;
+      margin-top: 1vh;
+    }
+
+    a {
+      display: flex;
+      border-radius: $borderRadiusInputs;
+      box-shadow: 10px 12px 15px rgba(0, 0, 0, 0.25);
+      height: 25%;
+      width: 90%;
+      margin: 2%;
+      background-color: white;
+      font-size: 2.5em;
+      text-decoration: none;
+      color: black;
+      text-align: center;
+      align-items: center;
+      justify-content: center;
+    }
+
+    a:hover {
+      background-color: #bab8b8;
+      color: black;
+    }
+  }
+}
