+++ src/Pages/dashboardPage/css/Dashboard.scss
@@ -1,0 +1,126 @@
+@import "../../../GlobalAssets/styles/constants.scss";
+
+.all-dashboard {
+  height: 100vh;
+  overflow-y: hidden;
+}
+
+.dashboard {
+  display: grid;
+  height: 88%;
+  grid-template-columns: 40% 40%;
+  grid-template-rows: 0.2fr 1fr 1fr;
+  color: white;
+  gap: 20px 15px;
+  justify-content: center;
+
+  h1 {
+    align-self: center;
+    grid-column: 1 / span 2;
+    grid-row: 1;
+    margin-top: 3%;
+    font-size: 4vh;
+
+    span {
+      font-size: 3vh;
+      font-weight: lighter;
+    }
+  }
+
+  .flashcard-statistics {
+    border-radius: $borderRadius;
+    background-color: $primaryMono1;
+    box-shadow: 2px 2px 10px $primaryTint10;
+  }
+
+  .flashcard-of-the-day {
+    border-radius: $borderRadius;
+    background-color: $primaryMono1;
+    box-shadow: 10px 12px 15px rgba(0, 0, 0, 0.35);
+  }
+
+  .flashcard-statistics {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    box-shadow: 10px 12px 15px rgba(0, 0, 0, 0.35);
+  }
+
+  .codewars-container {
+    border-radius: $borderRadius;
+    background-color: $primaryMono1;
+    padding-bottom: 3%;
+    box-shadow: 10px 12px 15px rgba(0, 0, 0, 0.35);
+
+    .form-header {
+      text-align: center;
+      font-size: 4.5vh;
+      margin: 5%;
+    }
+
+    .codewars-stats-container {
+      height: 100%;
+      display: flex;
+      flex-direction: column;
+
+      h2 {
+        margin-top: 7%;
+        text-align: center;
+        font-size: 4.5vh;
+      }
+      
+      h3 {
+        margin: 5%;
+        margin-bottom: 1%;
+        font-size: 3vh;
+        text-align: center;
+      }
+
+      li {
+        text-align: center;
+        font-size: 2vh;
+      }
+
+      p {
+        font-size: 2vh;
+        text-align: center;
+      }
+      ul {
+        list-style-type: none;
+      }
+    }
+
+    form {
+      align-items: center;
+      height: 100%;
+      display: flex;
+      flex-direction: column;
+      justify-content: space-evenly;
+
+      .cw-submit-button {
+        color: black;
+
+        &:hover {
+          background-color: #bab8b8;
+          cursor: pointer;
+        }
+      }
+
+      label {
+        color: white;
+        place-self: center;
+      }
+
+      input {
+        text-align: center;
+        border-radius: $borderRadiusInputs;
+        border: none;
+        box-shadow: 10px 12px 15px rgba(0, 0, 0, 0.25);
+        font-size: 3vh;
+        height: 5vh;
+        width: 80%;
+        place-self: center;
+      }
+    }
+  }
+}
