+++ src/Pages/flashcardPage /css/DeleteFlashCardButton.scss
