+++ src/Pages/flashcardPage /css/Flashcard.scss
@@ -1,0 +1,49 @@
+@import "../../../GlobalAssets/styles/constants.scss";
+
+.flashcard-container {
+  background-color: $primaryTint10;
+  border-radius: $borderRadius;
+  box-shadow: 0px 15px 15px rgba(0, 0, 0, 0.75);
+  color: white;
+  .flashcard-front,
+  .flashcard-back {
+    height: 50vh;
+    width: 55vw;
+    display: grid;
+    grid-template-columns: 1fr;
+    .rating {
+      justify-self: end;
+      align-items: center;
+      margin-top: 2%;
+      margin-right: 2%;
+      border-radius: $borderRadiusInputs;
+    }
+    .flashcard-footer {
+      grid-row: 4;
+      grid-column: 1;
+      justify-self: end;
+      align-self: end;
+      margin-right: 2%;
+      margin-bottom: 2%;
+    }
+    h2 {
+      font-size: 2em;
+      margin-left: 2%;
+      margin-bottom: 2%;
+    }
+    p {
+      margin-right: 2%;
+      margin-left: 2%;
+      height: 30vh;
+      font-size: 1em;
+    }
+    button {
+      justify-self: start;
+      grid-row: 4;
+      grid-column: 1;
+      margin-left: 2%;
+      margin-bottom: 2%;
+      font-size: 1em;
+    }
+  }
+}
