+++ src/Pages/flashcardPage /css/FlashcardCarousel.scss
@@ -1,0 +1,37 @@
+@import "../../../GlobalAssets/styles/constants.scss";
+
+.carousel-container {
+  height: 80vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin-bottom: 5%;
+
+  .swiper {
+    height: 60vh;
+    width: 80vw;
+    align-self: end;
+  }
+
+  .carousel-bottom-nav-container {
+    width: 80vw;
+    display: flex;
+    align-self: center;
+    justify-self: center;
+    justify-content: space-between;
+    margin: 0.5vw;
+    background-color: transparent;
+
+    button {
+      width: 30%;
+      box-shadow: 0px 10px 10px rgba(0, 0, 0, 0.5);
+    }
+  }
+}
+
+.flashcard {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
