+++ src/Pages/loginPage/css/Login.scss
@@ -1,0 +1,117 @@
+* {
+  margin: 0;
+  font-family: "Roboto", sans-serif;
+}
+
+.main-container {
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+
+  .login-container {
+    background-color: grey;
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    height: 50%;
+    width: 60%;
+    border-radius: 15px;
+    box-shadow: 10px 12px 15px rgba(0, 0, 0, 0.25);
+
+    .left-info-container {
+      display: flex;
+      flex-direction: column;
+      background-color: white;
+      box-shadow: 10px 12px 15px rgba(0, 0, 0, 0.25);
+      z-index: 5;
+      border-radius: 15px;
+      height: 100%;
+      width: 50%;
+
+      img {
+        height: 4vh;
+      }
+
+      h1 {
+        text-align: center;
+        margin-top: 12%;
+        margin-bottom: 5%;
+        font-size: 4vh;
+      }
+
+      p {
+        font-size: 2vh;
+        margin: 2%;
+        margin-left: 15%;
+      }
+    }
+
+    .right-side-container {
+      display: flex;
+      background-color: grey;
+      border-radius: 15px;
+      flex-direction: column;
+      justify-content: space-between;
+      align-items: center;
+      height: 100%;
+      width: 50%;
+
+      .ask-login {
+        font-size: 3vh;
+        margin-top: 12%;
+        margin-bottom: 12%;
+      }
+
+      .ask-signup {
+        margin: 12%;
+        font-size: 2.3vh;
+      }
+
+      .ask-signup:hover {
+        cursor: pointer;
+        color: white;
+      }
+
+      .login-error-message {
+        color: red;
+        font-size: 2vh;
+        text-align: center;
+      }
+
+      form {
+        height: 40%;
+        width: 80%;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        input {
+          margin-bottom: 10%;
+          font-size: 3vh;
+          text-align: center;
+          width: 100%;
+          height: 30%;
+          border-radius: 8px;
+          box-shadow: 0px 8px 15px rgba(0, 0, 0, 0.25);
+        }
+        .login-button {
+          width: 50%;
+          height: 30%;
+          border: solid 4px black;
+          border-radius: 30px;
+          box-shadow: 0px 8px 15px rgba(0, 0, 0, 0.25);
+          font-size: 3vh;
+
+          &:hover {
+            cursor: pointer;
+            background-color: #bab8b8;
+            color: white;
+          }
+        }
+      }
+    }
+  }
+}
